requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==4.0.39
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6